+++ ads-out.js
@@ -13,7 +13,10 @@ module.exports = function (RED) {
       this.on("input", function(msg) {
         debug('adsNotificationNode:','onAdsData:','node.id',node.id,'msg',msg)
         var cfg = {
+          useIndex: config.useIndex,
           symname: config.varName,
+          indexGroup: config.indexGroup,
+          indexOffset: config.indexOffset,
           adstype: config.varTyp,
           bytelength: config.varSize,
           timezone: config.timezone,
@@ -24,6 +27,15 @@ module.exports = function (RED) {
         if (typeof msg.config  !== 'undefined') {
           if (typeof msg.config.varName !== 'undefined') {
             cfg.symname = msg.config.varName
+          }
+          if (typeof msg.config.indexGroup !== 'undefined') {
+            cfg.indexGroup = msg.config.indexGroup
+          }
+          if (typeof msg.config.indexOffset !== 'undefined') {
+            cfg.indexOffset = msg.config.indexOffset
+          }
+          if (typeof msg.config.useIndex !== 'undefined') {
+            cfg.useIndex = msg.config.useIndex
           }
           if (typeof msg.config.varType !== 'undefined') {
             cfg.adstype = msg.config.varType
@@ -41,6 +53,22 @@ module.exports = function (RED) {
             cfg.topic = msg.config.topic||''
           }
         }
+
+        if (cfg.useIndex) {
+          delete(cfg.symname)
+          cfg.indexGroup = parseInt(cfg.indexGroup.toString())
+          if (isNaA(cfg.indexGroup)) {
+            cfg.indexGroup = 0
+          }
+          cfg.indexOffset = parseInt(cfg.indexOffset.toString())
+          if (isNaA(cfg.indexOffset)) {
+            cfg.indexOffset = 0
+          }
+        } else {
+          delete(cfg.indexGroup)
+          delete(cfg.indexOffset)
+        }
+
         cfg.hasTopic = String(cfg.topic).length > 0
         debug('adsNotificationNode:','onAdsData:','node.id',node.id,'cfg',cfg)
         var value = RED.util.getMessageProperty(msg,(cfg.outValue||'payload'))